requirements.txt
fastapi==0.135.3
uvicorn==0.41.0
pyngrok==8.0.0
python-dotenv==1.2.1
httpx==0.28.1
langsmith==0.7.26
langgraph>=0.2.0
mcp>=1.0.0
google-genai==1.70.0
opentelemetry-api==1.39.1
pytest==9.0.3
pytest-asyncio==1.3.0
anyio==4.12.1
anthropic==0.45.2
groq==0.18.0
